build: do not set get_option('prefix') to a variable
We seeing that variable, we don't know whether it's a global variable or
not. No need to use variables for everything, get_option('prefix') is
short enough and everybody understands it, while it's harder to remember
to use the 'prefix' variable, or when using that variable we need to
look where it is assigned to know what it is.