terminal: no longer use gedit_utils_menu_position_under_widget()
gedit_utils_menu_position_under_widget() and gtk_menu_popup() are
deprecated, so use gtk_menu_popup_at_pointer() and
gtk_menu_popup_at_widget() instead.

BTW, there was a runtime error with the lambda:
TypeError: <lambda>() takes 1 positional argument but 4 were given

The first case, with popup_at_pointer(), is triggerred with a
right-click event in the embedded terminal. The second case, with
popup_at_widget(), can be triggerred with the keyboard popup menu key.

For popup_at_widget(), the Gdk.Gravity values have been chosen because
the embedded terminal is located in the *bottom* panel of gedit, so it
makes more sense to show the popup menu just *above* the terminal, not
below.