drawspaces: remove menu item
The new GUI for this plugin will be:
- When the plugin is enabled, space drawing is enabled (bind the two).
- When the plugin is enabled, be able to access options: configure the
  space types to draw for each locations, roughly.
The two will be close to each other (like with every standard
libpeas-gtk plugin providing a configuration).

The problem with the menu item was that it was separate from the rest of
the configuration. It was a bit awkward, there were two ways to disable
space drawing: by disabling the plugin, and with the menu item. When the
menu item was disabled, another problem is that it could be hard to know
why the spaces are not drawn even though the plugin and the options are
enabled: three things needed to be enabled in order for the white spaces
to be drawn: the plugin needed to be enabled, the menu item too, and the
plugin configuration plays a role too (in case all checkbuttons are
disabled).

It now takes more clicks to disable space drawing, we need to go to the
preferences dialog and disable the plugin. But once the plugin is
correctly configured we probably don't often need to disable/enable it.

So I think the advantage (clearer way to configure space drawing)
outweighs the disadvantage (more clicks to enable/disable space drawing,
which I think is not often needed).